Moved p3d io ui and propoerty definitions
