+++ Arma3ObjectBuilder/io_p3d/ui_mat.py
@@ -125,15 +125,11 @@ classes = (
 )
 
 
-def register():
+def register_ui():
     for cls in classes:
         bpy.utils.register_class(cls)
-    
-    print("\t" + "UI: material properties")
 
 
-def unregister():
+def unregister_ui():
     for cls in reversed(classes):
         bpy.utils.unregister_class(cls)
-    
-    print("\t" + "UI: material properties")
