--- Arma3ObjectBuilder/props/__init__.py
@@ -1,3 +1,0 @@
-from . import material
-from . import object
-from . import scene
