--- Arma3ObjectBuilder/ui/__init__.py
@@ -1,3 +1,0 @@
-from . import import_export_p3d
-from . import props_material
-from . import props_object_mesh
